requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
numpy==1.23.5
pandas==2.2.2
scikit-learn==1.3.2
tensorflow==2.12.0